refactor some actions
