+++ packages/web/src/app/(app)/@sidebar/components/defaultSidebar/index.tsx
@@ -2,7 +2,8 @@ import { cookies } from "next/headers";
 import { auth } from "@/auth";
 import { HOME_VIEW_COOKIE_NAME } from "@/lib/constants";
 import { HomeView } from "@/hooks/useHomeView";
-import { getConnectionStats, getOrgAccountRequests } from "@/actions";
+import { getConnectionStats } from "@/actions";
+import { getOrgAccountRequests } from "@/features/userManagement/actions";
 import { isServiceError } from "@/lib/utils";
 import { ServiceErrorException } from "@/lib/serviceError";
 import { __unsafePrisma } from "@/prisma";
