+++ packages/web/src/app/(app)/settings/layout.tsx
@@ -3,7 +3,8 @@ import { Metadata } from "next"
 import { redirect } from "next/navigation";
 import { auth } from "@/auth";
 import { isServiceError } from "@/lib/utils";
-import { getConnectionStats, getOrgAccountRequests } from "@/actions";
+import { getConnectionStats } from "@/actions";
+import { getOrgAccountRequests } from "@/features/userManagement/actions";
 import { ServiceErrorException } from "@/lib/serviceError";
 import { OrgRole } from "@prisma/client";
 import { env } from "@sourcebot/shared";
