+++ packages/web/src/app/(app)/settings/members/components/inviteMemberCard.tsx
@@ -8,10 +8,10 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { useCallback, useState } from "react";
 import { z } from "zod";
-import { PlusCircleIcon, Loader2, AlertCircle, AlertTriangle } from "lucide-react";
+import { PlusCircleIcon, Loader2, AlertTriangle } from "lucide-react";
 import { OrgRole } from "@prisma/client";
 import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle } from "@/components/ui/alert-dialog";
-import { createInvites } from "@/actions";
+import { createInvites } from "@/features/userManagement/actions";
 import { isServiceError } from "@/lib/utils";
 import { useToast } from "@/components/hooks/use-toast";
 import { useRouter } from "next/navigation";
