+++ packages/web/src/app/invite/actions.ts
@@ -1,15 +1,15 @@
 "use server";
 
-import { isServiceError } from "@/lib/utils";
-import { notAuthenticated, notFound, orgNotFound, ServiceError } from "@/lib/serviceError";
-import { sew } from "@/middleware/sew";
+import { createAudit } from "@/ee/features/audit/audit";
 import { addUserToOrganization, orgHasAvailability } from "@/lib/authUtils";
-import { StatusCodes } from "http-status-codes";
+import { SINGLE_TENANT_ORG_ID } from "@/lib/constants";
 import { ErrorCode } from "@/lib/errorCodes";
+import { notAuthenticated, notFound, orgNotFound, ServiceError } from "@/lib/serviceError";
+import { isServiceError } from "@/lib/utils";
+import { sew } from "@/middleware/sew";
 import { getAuthenticatedUser } from "@/middleware/withAuth";
 import { __unsafePrisma } from "@/prisma";
-import { SINGLE_TENANT_ORG_ID } from "@/lib/constants";
-import { createAudit } from "@/ee/features/audit/audit";
+import { StatusCodes } from "http-status-codes";
 
 export const joinOrganization = async (inviteLinkId?: string) => sew(async () => {
     const authResult = await getAuthenticatedUser();
@@ -180,4 +180,3 @@ export const getInviteInfo = async (inviteId: string) => sew(async () => {
         }
     };
 });
-
